+++ src/inventorum.ebay/inventorum/ebay/apps/products/tests/test_core_products_sync.py
@@ -18,7 +18,7 @@ from inventorum.ebay.apps.products.models import EbayProductModel, EbayItemVaria
 from inventorum.ebay.apps.products.tasks import periodic_core_products_sync_task
 from inventorum.ebay.apps.products.tests.factories import EbayProductFactory, PublishedEbayItemFactory
 from inventorum.ebay.lib.celery import celery_test_case, get_anonymous_task_execution_context
-from inventorum.ebay.tests import IntegrationTest
+from inventorum.ebay.tests import StagingTestAccount, IntegrationTest
 from inventorum.ebay.tests.testcases import EbayAuthenticatedAPITestCase, UnitTestCase
 from mock import PropertyMock
 from rest_framework import status
@@ -35,14 +35,32 @@ class IntegrationTestPeriodicCoreProductsSync(EbayAuthenticatedAPITestCase, Core
         self.prepare_account_for_publishing(self.account)
 
     @celery_test_case()
+    @IntegrationTest.use_cassette("core_products_sync/core_products_sync_integration_no_images_test.yaml",
+                                  filter_query_parameters=['start_date'], record_mode="never")
+    def test_integration_with_no_images(self):
+        product_1_inv_id = self.create_core_api_product(
+            name="Test product 1",
+            description="Awesome test products are awesome",
+            gross_price="1.99",
+            quantity=12,
+            images=None,
+        )
+
+        response = self.client.get("/products/%s" % product_1_inv_id)
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.data)
+
+    @celery_test_case()
     @IntegrationTest.use_cassette("core_products_sync/core_products_sync_integration_test.yaml",
                                   filter_query_parameters=['start_date'], record_mode="never")
     def test_integration(self):
         # create some core products to play with
-        product_1_inv_id = self.create_core_api_product(name="Test product 1",
-                                                        description="Awesome test products are awesome",
-                                                        gross_price="1.99",
-                                                        quantity=12)
+        product_1_inv_id = self.create_core_api_product(
+            name="Test product 1",
+            description="Awesome test products are awesome",
+            gross_price="1.99",
+            quantity=12,
+            images=[{"id": StagingTestAccount.VALID_IMAGE_2_ID}, {"id": StagingTestAccount.VALID_IMAGE_ID}]
+        )
         ebay_product_1 = self.get_valid_ebay_product_for_publishing(self.account, inv_id=product_1_inv_id)
 
         product_2_inv_id = self.create_core_api_product(name="Test product 2",
@@ -139,10 +157,14 @@ class UnitTestCoreProductsSync(UnitTestCase):
         schedule_ebay_product_deletion = "inventorum.ebay.apps.products.tasks.schedule_ebay_product_deletion"
         self.schedule_ebay_product_deletion_mock = self.patch(schedule_ebay_product_deletion)
 
+        schedule_ebay_item_unpublish = "inventorum.ebay.apps.products.tasks.schedule_ebay_item_unpublish"
+        self.schedule_ebay_item_unpublish_mock = self.patch(schedule_ebay_item_unpublish)
+
     def reset_mocks(self):
         self.core_api_mock.reset_mock()
         self.schedule_ebay_item_update_mock.reset_mock()
         self.schedule_ebay_product_deletion_mock.reset_mock()
+        self.schedule_ebay_item_unpublish_mock.reset_mock()
 
     def expect_modified(self, *pages):
         mock = self.core_api_mock.get_paginated_product_delta_modified
@@ -308,7 +330,7 @@ class UnitTestCoreProductsSync(UnitTestCase):
                                         product=product_e,
                                         inv_product_id=1005)
 
-        assert all([p.is_published for p in [product_a, product_b, product_c, product_d, product_e]])
+        assert all(p.is_published for p in [product_a, product_b, product_c, product_d, product_e])
 
         self.expect_modified([delta_a], [delta_b], [delta_c])
         self.expect_deleted([1004, 1005])
@@ -331,7 +353,6 @@ class UnitTestCoreProductsSync(UnitTestCase):
         self.assertEqual(item_c_update.gross_price, D("111.11"))
         self.assertEqual(item_c_update.quantity, 22)
 
-        self.assertEqual(self.schedule_ebay_item_update_mock.call_count, 3)
         calls = self.schedule_ebay_item_update_mock.call_args_list
         self.assertEqual([args[0] for args, kwargs in calls], [item_a_update.id, item_b_update.id, item_c_update.id])
 
@@ -340,6 +361,35 @@ class UnitTestCoreProductsSync(UnitTestCase):
 
         calls = self.schedule_ebay_product_deletion_mock.call_args_list
         self.assertEqual([args[0] for args, kwargs in calls], [product_d.id, product_e.id])
+
+    def test_update_product_with_stock_in_zero_unpublishes_the_ebay_item(self):
+        # product that went out of stock
+        subject = CoreProductsSync(account=self.account)
+
+        product = EbayProductFactory.create(account=self.account)
+        item = PublishedEbayItemFactory.create(
+            account=self.account,
+            product=product,
+            inv_product_id=1006,
+            gross_price=D("5.64"),
+            quantity=79
+        )
+        delta = CoreProductDeltaFactory(id=1006, gross_price=D("3.99"), quantity=0)
+        self.assertTrue(product.is_published)
+
+        self.expect_modified([delta])
+        self.expect_deleted([])
+
+        subject.run()
+
+        self.assertEqual(item.updates.count(), 1)
+        update = item.updates.first()
+        self.assertEqual(update.gross_price, D('3.99'))
+        self.assertEqual(update.quantity, 0)
+
+        self.schedule_ebay_item_update_mock.assert_not_called()
+        calls = self.schedule_ebay_item_unpublish_mock.call_args_list
+        self.assertEqual([args[0] for args, kwargs in calls], [update.id])
 
     def test_published_modified_and_deleted_variations(self):
         subject = CoreProductsSync(account=self.account)