+++ src/inventorum.ebay/inventorum/ebay/lib/ebay/tests/test_ebay_items.py
@@ -1,11 +1,12 @@
 # encoding: utf-8
 from __future__ import absolute_import, unicode_literals
 import logging
-from decimal import *
+from decimal import Decimal
 
 from inventorum.ebay.tests import EbayTest
 from inventorum.ebay.lib.ebay import EbayConnectionException
-from inventorum.ebay.lib.ebay.data.items import EbayFixedPriceItem, EbayItemShippingService, EbayPicture
+from inventorum.ebay.lib.ebay.data.items import EbayFixedPriceItem, EbayItemShippingService, EbayPicture, \
+    EbayItemSpecific
 from inventorum.ebay.lib.ebay.items import EbayItems
 from inventorum.ebay.tests.testcases import EbayAuthenticatedAPITestCase
 
@@ -36,13 +37,13 @@ class TestEbayItems(EbayAuthenticatedAPITestCase):
         return EbayFixedPriceItem(
             sku="test_321",
             title="Inventorum iPad Stand",
-            description="<![CDATA[Stand for iPad from Inventorum for your POS Shop]]>",
+            description="Stand for iPad from Inventorum for your POS Shop",
             listing_duration="Days_120",
             country="DE",
             postal_code="13355",
             quantity=1,
             start_price=Decimal(599.99),
-            paypal_email_address="bartosz@hernas.pl",
+            paypal_email_address="info@inventorum.com",
             payment_methods=['PayPal'],
             pictures=[picture],
             category_id="64540",
@@ -69,52 +70,47 @@ class TestEbayItems(EbayAuthenticatedAPITestCase):
             postal_code="13355",
             quantity=1,
             start_price=Decimal(45.99),
-            paypal_email_address="bartosz@hernas.pl",
+            paypal_email_address="info@inventorum.com",
             payment_methods=['PayPal'],
             category_id="176973",
+            ean='9780471117094',
             shipping_services=[shipping],
-            pictures=[picture]
+            pictures=[picture],
+            item_specifics=[
+                EbayItemSpecific('Brand', ['Felt']),
+                EbayItemSpecific('Herstellernummer', ['34534jns']),
+            ]
         )
 
     @EbayTest.use_cassette("ebay_publish_ipad_stand_no_image.yaml")
     def test_failed_publishing(self):
         item = self._build_wrong_item()
         service = EbayItems(self.ebay_token)
+
         with self.assertRaises(EbayConnectionException) as e:
             service.publish(item)
 
         errors = e.exception.errors
-        self.assertEqual(len(errors), 4)
+        self.assertEqual(len(errors), 3)
 
-        self.assertEqual(errors[0].long_message, 'Für diese Kategorie ist kein Artikelzustand verfügbar. '
-                                                 'Der eingegebene Artikelzustand wurde entfernt.')
-        self.assertEqual(errors[0].short_message, 'Artikelzustand kann nicht verwendet werden.')
-        self.assertEqual(errors[0].code, 21917121)
-        self.assertEqual(errors[0].severity_code, 'Warning')
+        self.assertEqual(errors[0].long_message, 'Bei der ausgewählten Kategorie handelt es sich nicht um eine so '
+                                                 'genannte Unterkategorie.')
+        self.assertEqual(errors[0].code, 87)
+        self.assertEqual(errors[0].severity_code, 'Error')
         self.assertEqual(errors[0].classification, 'RequestError')
 
-        self.assertEqual(errors[1].long_message, 'Erforderliche Mindesanzahl an Bildern:  1 Für Angebote in dieser '
-                                                 'Kategorie empfehlen wir Ihnen, mindestens 2 Fotos hochzuladen, um '
-                                                 'Ihre Verkaufschancen  möglicherweise um 12 zu erhöhen. '
-                                                 '(Prozentangabe beruht auf Anteilen an verkauften Artikeln in dieser '
-                                                 'Kategorie mit unterschiedlich vielen Bildern. Tatsächliche '
-                                                 'Ergebnisse können anders ausfallen und der Verkauf ist nicht '
-                                                 'garantiert.)')
-        self.assertEqual(errors[1].code, 21919136)
-        self.assertEqual(errors[1].severity_code, 'Error')
-        self.assertEqual(errors[1].classification, 'RequestError')
-
-        self.assertEqual(errors[2].long_message, 'Bei der ausgewählten Kategorie handelt es sich nicht um eine so '
-                                                 'genannte Unterkategorie.')
-        self.assertEqual(errors[2].code, 87)
-        self.assertEqual(errors[2].severity_code, 'Error')
+        self.assertEqual(errors[2].long_message, 'Für diese Kategorie ist kein Artikelzustand verfügbar. '
+                                                 'Der eingegebene Artikelzustand wurde entfernt.')
+        self.assertEqual(errors[2].short_message, 'Artikelzustand kann nicht verwendet werden.')
+        self.assertEqual(errors[2].code, 21917121)
+        self.assertEqual(errors[2].severity_code, 'Warning')
         self.assertEqual(errors[2].classification, 'RequestError')
 
-        self.assertEqual(errors[3].long_message, 'Die Dauer "120" (in Tagen) ist für dieses Angebotsformat nicht '
+        self.assertEqual(errors[1].long_message, 'Die Dauer "120" (in Tagen) ist für dieses Angebotsformat nicht '
                                                  'verfügbar, bzw. ungültig für die Kategorie "64540".')
-        self.assertEqual(errors[3].code, 83)
-        self.assertEqual(errors[3].severity_code, 'Error')
-        self.assertEqual(errors[3].classification, 'RequestError')
+        self.assertEqual(errors[1].code, 83)
+        self.assertEqual(errors[1].severity_code, 'Error')
+        self.assertEqual(errors[1].classification, 'RequestError')
 
     @EbayTest.use_cassette("ebay_publish_ipad_stand_correct_then_unpublish_it.yaml")
     def test_publishing(self):
@@ -138,7 +134,5 @@ class TestEbayItems(EbayAuthenticatedAPITestCase):
         with self.assertRaises(EbayConnectionException) as e:
             service.publish(item)
 
-        self.assertTrue(e.exception.ebay_message)
+        self.assertIsNone(e.exception.ebay_message)
         self.assertTrue(e.exception.serialized_errors[0]['parameters'])
-        params = e.exception.serialized_errors[0]['parameters']
-        self.assertTrue(params[0].startswith('<div>'), 'Does not starts with <div>: %s' % params[0])